test: make LoanCalculation tests pass by changing double to decimal used for financial calculations

+++ 03-loan/LoanManagementApp/CSVWriter.cs
@@ -4,7 +4,7 @@ public class CSVWriter
 {
     private const string PAYMENT_HEADER = "Monthly Payment Number,Principal Paid,Remaining Balance";
         
-    public static void WriteToFile(List<Payment> payments, string filePath, double totalCost)
+    public static void WriteToFile(List<Payment> payments, string filePath, decimal totalCost)
     {
         using (StreamWriter writer = new StreamWriter(filePath))
         {
