+++ 03-loan/LoanManagementApp/Loan.cs
@@ -2,18 +2,18 @@
 
 public class Loan
 {
-    private double _principal;
+    private decimal _principal;
     private int _termInMonths;
-    private double _annualInterestRate;
+    private decimal _annualInterestRate;
 
-    public Loan(double principal, int termInMonths, double annualInterestRate)
+    public Loan(decimal principal, int termInMonths, decimal annualInterestRate)
     {
         _principal = principal;
         _termInMonths = termInMonths;
         _annualInterestRate = annualInterestRate;
     }
 
-    public double GetPrincipal()
+    public decimal GetPrincipal()
     {
         return _principal;
     }
@@ -23,7 +23,7 @@ public class Loan
         return _termInMonths;
     }
 
-    public double GetAnnualInterestRate()
+    public decimal GetAnnualInterestRate()
     {
         return _annualInterestRate;
     }
