+++ 03-loan/LoanManagementApp/LoanApp.cs
@@ -6,14 +6,14 @@
         {
             LoanInputHandler inputHandler = new LoanInputHandler(args);
 
-            double principal = inputHandler.GetPrincipal();
+            decimal principal = inputHandler.GetPrincipal();
             int termInMonths = inputHandler.GetTermInMonths();
-            double annualInterestRate = inputHandler.GetAnnualInterestRate();
+            decimal annualInterestRate = inputHandler.GetAnnualInterestRate();
 
             Loan loan = new Loan(principal, termInMonths, annualInterestRate);
             LoanCalculator calculator = new LoanCalculator(loan);
             List<Payment> payments = calculator.GenerateAmortizationSchedule();
-            double totalCost = calculator.CalculateTotalCreditCost();
+            decimal totalCost = calculator.CalculateTotalCreditCost();
 
             CSVWriter.WriteToFile(payments, "amortization_schedule.csv", totalCost);
         }
